server/requirements.txt
appdirs==1.4.4
blinker==1.4
braintree==4.1.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
distlib==0.3.0
dominate==2.5.1
ecdsa==0.15
filelock==3.0.12
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mock==4.0.2
pbr==5.4.5
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-dotenv==0.13.0
requests==2.23.0
sentry-sdk==0.14.4
six==1.14.0
SQLAlchemy==1.3.15
urllib3==1.25.8
virtualenv==20.0.20
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
